requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# OpenAI
openai>=1.0.0

# Anthropic (for Claude)
anthropic>=0.18.0

# LangChain for RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Vector Store
pymilvus>=2.3.0

# Document Processing
pypdf>=4.0.0

# Google Sheets Integration
gspread>=5.12.0
google-auth>=2.25.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
